\name{GWsignif-package}
\alias{GWsignif-package}
\docType{package}
\title{
Estimating Genome-Wide Significance for Whole Genome Sequencing Studies, Either Single SNP Tests or Region-Based Tests
}
\description{
The correlations and linkage disequilibrium between tests can vary as a function of minor allele frequency thresholds used to filter variants, and also varies with different choices of test statistic for region-based tests. Appropriate genome-wide significance thresholds can be estimated empirically through permutation on only a small proportion of the whole genome. 
}
\details{
\tabular{ll}{
Package: \tab GWsignif\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-06-19\cr
License: \tab GLP 2.0 or greater\cr
}
}
\author{
ChangJiang Xu and Celia M.T. Greenwood
Maintainer: ChangJiang Xu <changjiang.h.xu@gmail.com>
}
\references{
ChangJiang Xu, Ioanna Tachmazidou, Klaudia Walter, Antonio Ciampi, Eleftheria Zeggini, Celia M.T. Greenwood (2014) Estimating genome-wide significance for whole genome sequencing studies. Genetic Epidemiology
}
\keyword{ package }
