\name{GWsignif-package}
\alias{GWsignif-package}
\docType{package}
\title{
Genome-wide significance for whole genome sequencing studies
}
\description{
Estimate genome-wide significance using extrapolation approach: estimate significance thresholds in small regions and then extrapolate a large region significance threshold.
}
\details{
\tabular{ll}{
Package: \tab GWsignif\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-14\cr
License: \tab GLP 2.0 or greater\cr
}
}
\author{
ChangJiang Xu and Celia M.T. Greenwood
Maintainer: ChangJiang Xu <changjiang.h.xu@gmail.com>
}
\references{
ChangJiang Xu, Ioanna Tachmazidou, Klaudia Walter, Antonio Ciampi, Eleftheria Zeggini, Celia M.T. Greenwood (2014) Estimating genome-wide significance for whole genome sequencing studies. Genetic Epidemiology
}
\keyword{ package }
